+++ api/grpcserver/node_service.go
@@ -76,9 +76,9 @@ func (s NodeService) Status(ctx context.Context, request *pb.StatusRequest) (*pb
 		Status: &pb.NodeStatus{
 			ConnectedPeers: s.PeerCounter.PeerCount(),            // number of connected peers
 			IsSynced:       s.Syncer.IsSynced(),                  // whether the node is synced
-			SyncedLayer:    s.Tx.LatestLayer().Uint64(),          // latest layer we saw from the network
-			TopLayer:       s.GenTime.GetCurrentLayer().Uint64(), // current layer, based on time
-			VerifiedLayer:  s.Tx.LatestLayerInState().Uint64(),   // latest verified layer
+			SyncedLayer:    s.Tx.LatestLayer().Uint32(),          // latest layer we saw from the network
+			TopLayer:       s.GenTime.GetCurrentLayer().Uint32(), // current layer, based on time
+			VerifiedLayer:  s.Tx.LatestLayerInState().Uint32(),   // latest verified layer
 		},
 	}, nil
 }
