+++ events/events.go
@@ -82,7 +82,7 @@ func (p *EventPublisher) Close() error {
 // NewBlock is sent when a new block is created by this miner
 type NewBlock struct {
 	ID    string
-	Layer uint64
+	Layer uint32
 	Atx   string
 }
 
@@ -94,7 +94,7 @@ func (NewBlock) GetChannel() ChannelID {
 // DoneCreatingBlock signals that this miner has created a block
 type DoneCreatingBlock struct {
 	Eligible bool
-	Layer    uint64
+	Layer    uint32
 	Error    string
 }
 
@@ -117,7 +117,7 @@ func (ValidBlock) GetChannel() ChannelID {
 // NewAtx signals that a new ATX has been received
 type NewAtx struct {
 	ID      string
-	LayerID uint64
+	LayerID uint32
 }
 
 // GetChannel gets the message type which means on which this message should be sent
@@ -176,7 +176,7 @@ func (RewardReceived) GetChannel() ChannelID {
 type AtxCreated struct {
 	Created bool
 	ID      string
-	Layer   uint64
+	Layer   uint32
 }
 
 // GetChannel gets the message type which means on which this message should be sent
